Fixing eslint errors
